NEW: statut session, brouillon/validé

+++ session_calendrier.php
@@ -41,61 +41,69 @@ if (! empty ( $id )) {
 
 switch ($action) {
 	case 'addtimeslot' :
-		$date = GETPOST('date', 'alpha');
-		$heure_debut = GETPOST('heure_debut', 'alpha');
-		$heure_fin = GETPOST('heure_fin', 'alpha');
-
-		if(strcmp($heure_debut, $heure_fin) < 0) {
-			$TDate = explode('/', $date);
-
-			$dateSQL = $TDate[2] . '-' . $TDate[1] . '-' . $TDate[0] . ' ';
-			$date_debut = $dateSQL.$heure_debut.':00';
-			$date_fin = $dateSQL.$heure_fin.':00';
-
-			if(! $session->hasCreneauxBetween($PDOdb, $date_debut, $date_fin)) {
-				$THeureDebut = explode(':', $heure_debut);
-				$THeureFin = explode(':', $heure_fin);
-
-				$creneau->fk_session = $session->rowid;
-				$creneau->debut = dol_mktime($THeureDebut[0], $THeureDebut[1], 0, $TDate[1], $TDate[0], $TDate[2]);
-				$creneau->fin = dol_mktime($THeureFin[0], $THeureFin[1], 0, $TDate[1], $TDate[0], $TDate[2]);
-
-				if($creneau->debut >= $session->date_debut && $creneau->fin <= ($session->date_fin + 86400)) { // date_fin -> le jour de la fin à minuit, on rajoute un jour en rajoutant 86400s
-					$dureeHeures = ($creneau->fin - $creneau->debut) / 3600;
+		if($session->statut == 0) {
+			$date = GETPOST('date', 'alpha');
+			$heure_debut = GETPOST('heure_debut', 'alpha');
+			$heure_fin = GETPOST('heure_fin', 'alpha');
 	
-					$session->duree_planifiee += $dureeHeures;
+			if(strcmp($heure_debut, $heure_fin) < 0) {
+				$TDate = explode('/', $date);
 	
-					if($session->duree_planifiee <= $formation->duree) {
-						$session->save($PDOdb);
-						$creneau->save($PDOdb);
+				$dateSQL = $TDate[2] . '-' . $TDate[1] . '-' . $TDate[0] . ' ';
+				$date_debut = $dateSQL.$heure_debut.':00';
+				$date_fin = $dateSQL.$heure_fin.':00';
+	
+				if(! $session->hasCreneauxBetween($PDOdb, $date_debut, $date_fin)) {
+					$THeureDebut = explode(':', $heure_debut);
+					$THeureFin = explode(':', $heure_fin);
+	
+					$creneau->fk_session = $session->rowid;
+					$creneau->debut = dol_mktime($THeureDebut[0], $THeureDebut[1], 0, $TDate[1], $TDate[0], $TDate[2]);
+					$creneau->fin = dol_mktime($THeureFin[0], $THeureFin[1], 0, $TDate[1], $TDate[0], $TDate[2]);
+	
+					if($creneau->debut >= $session->date_debut && $creneau->fin <= ($session->date_fin + 86400)) { // date_fin -> le jour de la fin à minuit, on rajoute un jour en rajoutant 86400s
+						$dureeHeures = ($creneau->fin - $creneau->debut) / 3600;
+		
+						$session->duree_planifiee += $dureeHeures;
+		
+						if($session->duree_planifiee <= $formation->duree) {
+							$session->save($PDOdb);
+							$creneau->save($PDOdb);
+						} else {
+							setEventMessage($langs->trans('PFSessionTooMuchTimePlanned'), 'errors');
+						}
 					} else {
-						setEventMessage($langs->trans('PFSessionTooMuchTimePlanned'), 'errors');
+						setEventMessage($langs->trans('PFTimeSlotOutOfSessionDates'), 'errors');
 					}
 				} else {
-					setEventMessage($langs->trans('PFTimeSlotOutOfSessionDates'), 'errors');
+					setEventMessage($langs->trans('PFTimeSlotOverlap'), 'errors');
 				}
 			} else {
-				setEventMessage($langs->trans('PFTimeSlotOverlap'), 'errors');
+				setEventMessage($langs->trans('PFStartTimeMustBeBeforeEndTime'), 'errors');
 			}
 		} else {
-			setEventMessage($langs->trans('PFStartTimeMustBeBeforeEndTime'), 'errors');
+			setEventMessage($langs->trans('PFTryingToEditAValidatedSession'), 'errors');
 		}
 
 		_list ( $PDOdb, $session, $formation, $creneau );
 	break;
 
 	case 'deletetimeslot' :
-		$timeslotRowid = GETPOST ( 'timeslot' );
-		
-		if ($timeslotRowid > 0) {
-			$creneau->load($PDOdb, $timeslotRowid);
-
-			$session->duree_planifiee -= ($creneau->fin - $creneau->debut) / 3600;
-
-			$creneau->delete($PDOdb);
-			$session->save($PDOdb);
+		if($session->statut == 0) {
+			$timeslotRowid = GETPOST ( 'timeslot' );
+			
+			if ($timeslotRowid > 0) {
+				$creneau->load($PDOdb, $timeslotRowid);
+	
+				$session->duree_planifiee -= ($creneau->fin - $creneau->debut) / 3600;
+	
+				$creneau->delete($PDOdb);
+				$session->save($PDOdb);
+			}
+		} else {
+			setEventMessage($langs->trans('PFTryingToEditAValidatedSession'), 'errors');
 		}
-		
+
 		_list ( $PDOdb, $session, $formation, $creneau );
 	break;
 
@@ -130,9 +138,9 @@ function _list(&$PDOdb, &$session, &$formation, &$creneau) {
 	foreach ( $TCreneaux AS $c ) {
 		$actionsButtons = '';
 
-		// if($pf->statut == 0) {
-		$actionButtons = '<a href="' . $_SERVER ['PHP_SELF'] . '?id=' . $session->id . '&action=deletetimeslot&timeslot=' . $c->rowid . '">' . img_picto ( '', 'delete' ) . '</a>';
-		// }
+		if($session->statut == 0) {
+			$actionButtons = '<a href="' . $_SERVER ['PHP_SELF'] . '?id=' . $session->id . '&action=deletetimeslot&timeslot=' . $c->rowid . '">' . img_picto ( '', 'delete' ) . '</a>';
+		}
 
 		$date_debut = dol_stringtotime(str_replace(' ', 'T', $c->debut), 0);
 		$date_fin = dol_stringtotime(str_replace(' ', 'T', $c->fin), 0);
@@ -147,8 +155,6 @@ function _list(&$PDOdb, &$session, &$formation, &$creneau) {
 		print '<td>' . secondesToHHMM($duree). '</td>';
 		print '<td align="right">' . $actionButtons . '</td>';
 		print '</tr>';
-
-		// }
 	}
 
 	if ($nbCreneaux == 0) {
@@ -168,31 +174,34 @@ function _list(&$PDOdb, &$session, &$formation, &$creneau) {
 
 	// Ajout nouveau créneau
 
-	print '<tr class="liste_titre"><td colspan="5">' . $langs->trans ( 'PFAddNewSessionTimeSlot' ) . '</td></tr>';
+	if($session->statut == 0) {
 
-	if($session->duree_planifiee < $formation->duree) {
-		print '<tr class="liste_titre">';
-		print '<th class="liste_titre">' . $langs->trans ( 'Date' ) . '</th>';
-		print '<th class="liste_titre">' . $langs->trans ( 'PFStartTime' ) . '</th>';
-		print '<th class="liste_titre">' . $langs->trans ( 'PFEndTime' ) . '</th>';
-		print '<th class="liste_titre">' . '' . '</th>';
-		print '<th class="liste_titre">&nbsp;</th>';
-		print '</tr>';
+		print '<tr class="liste_titre"><td colspan="5">' . $langs->trans ( 'PFAddNewSessionTimeSlot' ) . '</td></tr>';
 	
-		$formCore = new TFormCore ( $_SERVER ['PHP_SELF'] . '?id=' . $session->id, 'formAddAttendee', 'POST' );
-	
-		print '<tr class="impair">';
-		print '<td>' . $formCore->calendrier('', 'date', '') . '</td>';
-		print '<td>' . $formCore->timepicker('', 'heure_debut', '') . '</td>';
-		print '<td>' . $formCore->timepicker('', 'heure_fin', ''). '</td>';
-		print '<td>' . '' . '</td>';
-		print '<td align="right">' . $formCore->hidden ( 'action', 'addtimeslot' ) . $formCore->hidden ( 'fk_session', $session->rowid ) . $formCore->btsubmit ( $langs->trans ( 'Add' ), 'addtimeslot' ) . '</td>';
-		print '</tr>';
-	
-		$formCore->end();
-	
-	} else {
-		print '<tr><td colspan="5" align="center">' . $langs->trans ( 'PFFormationTimePlanned' ) . '</td></tr>';
+		if($session->duree_planifiee < $formation->duree) {
+			print '<tr class="liste_titre">';
+			print '<th class="liste_titre">' . $langs->trans ( 'Date' ) . '</th>';
+			print '<th class="liste_titre">' . $langs->trans ( 'PFStartTime' ) . '</th>';
+			print '<th class="liste_titre">' . $langs->trans ( 'PFEndTime' ) . '</th>';
+			print '<th class="liste_titre">' . '' . '</th>';
+			print '<th class="liste_titre">&nbsp;</th>';
+			print '</tr>';
+		
+			$formCore = new TFormCore ( $_SERVER ['PHP_SELF'] . '?id=' . $session->id, 'formAddAttendee', 'POST' );
+		
+			print '<tr class="impair">';
+			print '<td>' . $formCore->calendrier('', 'date', '') . '</td>';
+			print '<td>' . $formCore->timepicker('', 'heure_debut', '') . '</td>';
+			print '<td>' . $formCore->timepicker('', 'heure_fin', ''). '</td>';
+			print '<td>' . '' . '</td>';
+			print '<td align="right">' . $formCore->hidden ( 'action', 'addtimeslot' ) . $formCore->hidden ( 'fk_session', $session->rowid ) . $formCore->btsubmit ( $langs->trans ( 'Add' ), 'addtimeslot' ) . '</td>';
+			print '</tr>';
+		
+			$formCore->end();
+		
+		} else {
+			print '<tr><td colspan="5" align="center">' . $langs->trans ( 'PFFormationTimePlanned' ) . '</td></tr>';
+		}
 	}
 	
 	print "</table>";
